feat: useTermForm으로 분리
