+++ src/page/signUp/index/TermPage.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import Button from '@/common/component/Button/Button';
@@ -7,6 +6,7 @@ import Heading from '@/common/component/Heading/Heading';
 
 import TermItem from '@/page/signUp/index/component/TermItem/TermItem';
 import TermsAgreeButton from '@/page/signUp/index/component/TermsAgreeButton/TermsAgreeButton';
+import { useTermForm } from '@/page/signUp/index/hook/useTermForm';
 import { pageStyle } from '@/page/signUp/info/InfoFormPage.style';
 import { formStyle } from '@/page/signUp/info/component/InfoForm/InfoForm.style';
 
@@ -14,51 +14,19 @@ import { PATH } from '@/shared/constant/path';
 
 import { PERSONAL, TERM } from '@/mock/data/term';
 
-type TermItem = {
-  serviceTerm: boolean;
-  privatePolicy: boolean;
-  personalInfo: boolean;
-};
-
 const TermPage = () => {
-  const [totalAgreeClicked, setTotalAgreeClicked] = useState(false);
-  const [termStatus, setTermStatus] = useState<TermItem>({
-    serviceTerm: false,
-    privatePolicy: false,
-    personalInfo: false,
-  });
+  const { totalAgreeClicked, termStatus, isConfirmed, handleAllTermsAgree, handleTermAgree } = useTermForm();
 
   const navigate = useNavigate();
 
-  const isConfirmed = termStatus.serviceTerm && termStatus.privatePolicy;
-
-  const 약관전체동의클릭 = () => {
-    setTotalAgreeClicked((prev) => !prev);
-
-    setTermStatus({
-      serviceTerm: totalAgreeClicked ? false : true,
-      privatePolicy: totalAgreeClicked ? false : true,
-      personalInfo: totalAgreeClicked ? false : true,
-    });
-  };
-
-  const handleItemClick = (key: keyof TermItem) => {
-    setTermStatus((prev) => ({
-      ...prev,
-      [key]: !prev[key],
-    }));
-  };
-
-  const handleNextStep = () => {
-    navigate(PATH.SIGNUP_UNIV);
-  };
+  const handleNextStep = () => navigate(PATH.SIGNUP_UNIV);
 
   return (
     <Flex tag="main" css={pageStyle}>
       <form css={formStyle}>
         <Heading tag="H4">이용 약관 동의</Heading>
         <Flex styles={{ direction: 'column', width: '100%', gap: '1.6rem' }}>
-          <TermsAgreeButton isClicked={totalAgreeClicked} onClick={약관전체동의클릭} />
+          <TermsAgreeButton isClicked={totalAgreeClicked} onClick={handleAllTermsAgree} />
 
           <TermItem
             term="이용 약관"
@@ -66,7 +34,7 @@ const TermPage = () => {
             description="티키 서비스 이용약관은 다음과 같은 내용을 담고 있습니다."
             isRequired
             isSelected={termStatus.serviceTerm}
-            onSelect={() => handleItemClick('serviceTerm')}
+            onSelect={() => handleTermAgree('serviceTerm')}
           />
           <TermItem
             term="개인정보 처리 방침"
@@ -74,17 +42,17 @@ const TermPage = () => {
             description="티키 서비스 이용약관은 다음과 같은 내용을 담고 있습니다."
             isRequired
             isSelected={termStatus.privatePolicy}
-            onSelect={() => handleItemClick('privatePolicy')}
+            onSelect={() => handleTermAgree('privatePolicy')}
           />
           <TermItem
             term="개인정보 수집 및 이용"
             content={PERSONAL}
             description="티키 서비스 이용약관은 다음과 같은 내용을 담고 있습니다."
             isSelected={termStatus.personalInfo}
-            onSelect={() => handleItemClick('personalInfo')}
+            onSelect={() => handleTermAgree('personalInfo')}
           />
         </Flex>
-        <Button disabled={!isConfirmed} onClick={handleNextStep} variant="primary" size="large">
+        <Button disabled={!isConfirmed} onClick={handleNextStep} variant="primary" size="xLarge">
           다음
         </Button>
       </form>
